--- client/.eslintrc.json
@@ -1,32 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "es2021": true
-    },
-    "extends": [
-        "eslint:recommended",
-	  "plugin:react/recommended",
-	  "plugin:@typescript-eslint/recommended",
-	  "prettier" // Add this line!
-    ],
-    "overrides": [
-    ],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": "latest",
-        "sourceType": "module"
-    },
-    "plugins": [
-        "react", "react-hooks", "@typescript-eslint", "prettier"
-    ],
-    "rules": {
-        "react/react-in-jsx-scope": "off",
-        "camelcase": "error",
-        "no-duplicate-imports": "error"
-    },
-    "settings": {
-        "import/resolver": {
-        "typescript": {}
-        }
-    }
-}
