--- client/.prettierignore
@@ -1,4 +1,0 @@
-node_modules
-# Ignore artifacts:
-build
-coverage
