--- client/.prettierrc.json
@@ -1,9 +1,0 @@
-{
-    "semi": false,
-    "tabWidth": 2,
-    "printWidth": 100,
-    "singleQuote": true,
-    "trailingComma": "all",
-    "jsxSingleQuote": true,
-    "bracketSpacing": true
-}
