+++ client/package.json
@@ -3,19 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.10.3",
+    "@mui/material": "^5.10.5",
     "@progress/kendo-data-query": "^1.6.0",
     "@progress/kendo-drawing": "^1.17.1",
     "@progress/kendo-licensing": "^1.2.2",
-    "@progress/kendo-react-animation": "^5.6.0",
-    "@progress/kendo-react-buttons": "^5.6.0",
-    "@progress/kendo-react-data-tools": "^5.6.0",
-    "@progress/kendo-react-dateinputs": "^5.6.0",
-    "@progress/kendo-react-dropdowns": "^5.6.0",
-    "@progress/kendo-react-grid": "^5.6.0",
-    "@progress/kendo-react-inputs": "^5.6.0",
-    "@progress/kendo-react-intl": "^5.6.0",
-    "@progress/kendo-react-treeview": "^5.6.0",
-    "@progress/kendo-theme-material": "^5.8.1",
+    "@progress/kendo-react-animation": "^5.5.0",
+    "@progress/kendo-react-buttons": "^5.5.0",
+    "@progress/kendo-react-data-tools": "^5.5.0",
+    "@progress/kendo-react-dateinputs": "^5.5.0",
+    "@progress/kendo-react-dropdowns": "^5.5.0",
+    "@progress/kendo-react-grid": "^5.5.0",
+    "@progress/kendo-react-inputs": "^5.5.0",
+    "@progress/kendo-react-intl": "^5.5.0",
+    "@progress/kendo-react-treeview": "^5.5.0",
+    "@progress/kendo-react-upload": "^5.5.0",
+    "@progress/kendo-theme-default": "^5.6.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -26,6 +31,7 @@
     "axios": "^0.27.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1",
     "socket.io-client": "^4.5.2",
     "typescript": "^4.8.3",
@@ -35,10 +41,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint src/**/*.{js,jsx,ts,tsx,json}",
-    "lint:fix": "eslint --fix 'src/**/*.{js,jsx,ts,tsx,json}'",
-    "format": "prettier --write 'src/**/*.{js,jsx,ts,tsx,css,md,json}' --config ./.prettierrc"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -57,15 +60,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.37.0",
-    "@typescript-eslint/parser": "^5.37.0",
-    "eslint": "^8.23.1",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.8",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.7.1"
   }
 }
