+++ client/src/components/pages/landing.css
@@ -1,0 +1,8 @@
+.landing{
+    width: 100%;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
